Some refactoring
* Results in cleaner godocs for client sdk, exposing only public stuff
* consolidates crypto and attestation stuff all into corresponding
  (sub-) modules inside internal
