+++ internal/crypto/crypto.go
@@ -5,7 +5,7 @@ Copyright 2020 Intel Corporation
 SPDX-License-Identifier: Apache-2.0
 */
 
-package utils
+package crypto
 
 import (
 	"fmt"
